+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 10.7.1
+- Fixes an issue, where the Google Pay button was always displayed in English instead of the sales channel language (shopware/shopware#17804)
+
 # 10.7.0
 - Added merchant-configurable appearance settings for the Pay Later installment banner: logo type, text colour, and text size